# docs/swagger.yaml
basePath: /
definitions:
  controller.Biodata:
    properties:
      _id:
        type: string
      alamat:
        type: string
      doctor:
        $ref: '#/definitions/controller.Doctor'
      gender:
        type: string
      medicalRecord:
        $ref: '#/definitions/controller.MedicalRecord'
      pasienName:
        type: string
      phonenumber:
        type: string
      usia:
        type: string
    type: object
  controller.Doctor:
    properties:
      contact:
        type: string
      name:
        type: string
      specialty:
        type: string
    type: object
  controller.MedicalRecord:
    properties:
      _id:
        type: string
      diagnosis:
        type: string
      treatment:
        type: string
      visitdate:
        type: string
    type: object
  controller.ReqPasien:
    properties:
      alamat:
        example: Sariasih 25, Bandung
        type: string
      doctor:
        $ref: '#/definitions/controller.Doctor'
      gender:
        example: Laki-laki
        type: string
      medicalRecord:
        $ref: '#/definitions/controller.MedicalRecord'
      pasienName:
        example: Doni
        type: string
      phonenumber:
        example: "08567432"
        type: string
      usia:
        example: 20 Tahun
        type: string
    type: object
host: ws-rumahsakit-2e1eb71f14e2.herokuapp.com
info:
  contact:
    email: 714220040@std.ulbi.ac.id
    name: API Support
    url: https://github.com/Nidasakinaa
  description: This is a sample swagger for Fiber
  title: TES SWAGGER PASIEN
  version: "1.0"
paths:
  /delete/{id}:
    delete:
      consumes:
      - application/json
      description: Hapus data pasien.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Delete data pasien.
      tags:
      - Pasien
  /insert:
    post:
      consumes:
      - application/json
      description: Input data pasien.
      parameters:
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqPasien'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Biodata'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Insert data pasien.
      tags:
      - Pasien
  /pasien:
    get:
      consumes:
      - application/json
      description: Mengambil semua data pasien.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Biodata'
      summary: Get All Data Pasien.
      tags:
      - Pasien
  /pasien/{id}:
    get:
      consumes:
      - application/json
      description: Ambil per ID data pasien.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Biodata'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get By ID Data Pasien.
      tags:
      - Pasien
  /update/{id}:
    put:
      consumes:
      - application/json
      description: Ubah data pasien.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqPasien'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Biodata'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Update data pasien.
      tags:
      - Pasien
schemes:
- https
- http
swagger: "2.0"